backend/requirements.txt
click==7.1.2
dnspython==1.16.0
email-validator==1.1.2
eventlet==0.30.1
Flask==1.1.2
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-SocketIO==4.3.2
Flask-WTF==0.14.3
greenlet==1.0.0
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mongoengine==0.22.1
passlib==1.7.4
PyJWT==2.0.1
pymongo==3.11.3
python-engineio==3.14.2
python-socketio==4.6.1
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3
